update sensory motor bar coding analyses

+++ Notebooks/sensory_motor_bar_code.py
@@ -1,4 +1,4 @@
-from single_cell_class import *
+from sensory_motor_single_cell_class import *
 from utils import *
 from brain_seg import brain_seg_factor
 from factor import thres_factor_
@@ -138,8 +138,9 @@ def sensory_motor_bar_code(row):
             mlt_nopulse_trial.append(trial)
             mlt_nopulse_type.append(epoch_frame[trial]//5)
     
-    cell_msensory_stats = dFF_.map_blocks(multi_pulse_stats_chunks, pulse_trial=mlt_pulse_trial, nopulse_trial=mlt_nopulse_trial, t_pre=t_pre, t_post=t_post, dtype='O').compute() 
-    np.savez(save_root+'cell_type_stats_msensory', cell_msensory_stats=cell_msensory_stats)
+    if not os.path.exists(save_root+'cell_type_stats_msensory.npz'):
+        cell_msensory_stats = dFF_.map_blocks(multi_pulse_stats_chunks, pulse_trial=mlt_pulse_trial, nopulse_trial=mlt_nopulse_trial, t_pre=t_pre, t_post=t_post, dtype='O').compute() 
+        np.savez(save_root+'cell_type_stats_msensory', cell_msensory_stats=cell_msensory_stats)
     
     
     ###################################
@@ -172,15 +173,15 @@ def sensory_motor_bar_code(row):
                 nopulse_trial.append(trial+m*trial_len_-2)
                 nopulse_type.append(epoch_frame[trial+m*trial_len_]//5)
 
-    if len(pulse_trial)>0:
+    if (not os.path.exists(save_root+'cell_type_stats_sensory.npz')) and (len(pulse_trial)>0):
         cell_sensory_stats = dFF_.map_blocks(pulse_stats_chunks, pulse_trial=pulse_trial, nopulse_trial=nopulse_trial, dtype='O').compute() 
         np.savez(save_root+'cell_type_stats_sensory', cell_sensory_stats=cell_sensory_stats)
 
-    if (len(pulse_motor_trial)>0) and (probe_gain==0):
+    if (not os.path.exists(save_root+'cell_type_stats_pulse_motor.npz')) and (len(pulse_motor_trial)>0) and (probe_gain==0):
         cell_pulse_motor_stats = dFF_.map_blocks(pulse_stats_chunks, pulse_trial=pulse_motor_trial, nopulse_trial=nopulse_trial, dtype='O').compute()
         np.savez(save_root+'cell_type_stats_pulse_motor', cell_pulse_motor_stats=cell_pulse_motor_stats)
 
-    if (len(pulse_trial)>0) and (len(pulse_motor_trial)>0) and (probe_gain==0):
+    if (not os.path.exists(save_root+'cell_comp_pulse_motor_stats_.npz')) and (len(pulse_trial)>0) and (len(pulse_motor_trial)>0) and (probe_gain==0):
         cell_comp_pulse_motor_stats = dFF_.map_blocks(comp_stats_chunks, cond_trial=pulse_trial, comp_trial=pulse_motor_trial, pre=2, post=5, dtype='O').compute()  
         np.savez(save_root+'cell_comp_pulse_motor_stats_', cell_comp_pulse_motor_stats=cell_comp_pulse_motor_stats)
 
@@ -244,154 +245,15 @@ def sensory_motor_bar_code(row):
             noswim_trial.append(on_+off_+off_set)
 
     swim_trial_ = np.array(swim_trial)[np.array(swim_type)!=6]
-    if len(swim_trial_)>0:
+    if (not os.path.exists(save_root+'cell_type_stats_sm.npz')) and len(swim_trial_)>0:
         cell_sm_stats = dFF_.map_blocks(motor_stats_chunks, swim_trial=swim_trial_, noswim_trial=noswim_trial, swim_len=swim_len, pre_len=pre_len, dtype='O').compute()
         np.savez(save_root+'cell_type_stats_sm', cell_sm_stats=cell_sm_stats)
 
     swim_trial_ = np.array(swim_trial)[np.array(swim_type)==6]
-    if len(swim_trial_)>0:
+    if (not os.path.exists(save_root+'cell_type_stats_motor.npz')) and len(swim_trial_)>0:
         cell_motor_stats = dFF_.map_blocks(motor_stats_chunks, swim_trial=swim_trial_, noswim_trial=noswim_trial, swim_len=swim_len, pre_len=pre_len, dtype='O').compute() 
         np.savez(save_root+'cell_type_stats_motor', cell_motor_stats=cell_motor_stats)
         
     fdask.terminate_workers(cluster, client)
     return None
 
-
-def brain_state_bar_code(row):
-
-    save_root = row['save_dir']+'/'
-    # check ephys data
-    dat_dir = row['dat_dir'].replace('/im/', '/')
-    dat_dir = dat_dir.replace('/im_CM0/', '/')
-    dat_dir = dat_dir.replace('/im_CM1/', '/')
-    p_dir = dat_dir + 'processed/'
-    ephys_dir = dat_dir + 'ephys/'
-    if not os.path.exists(ephys_dir):
-        print('Missing directory')
-        print(ephys_dir)
-        return None
-    
-    _ = np.load(save_root+'cell_dff.npz', allow_pickle=True)
-    A = _['A']
-    A_loc = _['A_loc']
-    dFF = _['dFF'].astype('float')
-    _ = None
-
-    brain_map = np.load(save_root+'Y_ave.npy').astype('float').squeeze()
-    A_center = np.load(save_root+'cell_center.npy')
-    A_center_grid = np.round(A_center).astype('int')
-    cells_in_mask = []
-
-    for n_layer in range(brain_map.shape[0]):
-        layer_ = A_center[:, 0]==n_layer
-        cell_ids = np.where(layer_)[0]
-        mask_ = brain_map[n_layer]>2
-        y = A_center_grid[cell_ids, 2]
-        x = A_center_grid[cell_ids, 1]
-        x_max, y_max = mask_.shape
-        num_cells = len(cell_ids)
-        in_mask_ = np.zeros(num_cells).astype('bool')
-        for n in range(num_cells):
-            if (x[n]<x_max) and (y[n]<y_max):
-                in_mask_[n] = mask_[x[n], y[n]]
-        cells_in_mask.append(cell_ids[in_mask_])
-    cells_in_mask = np.concatenate(cells_in_mask)
-    A_center = A_center[cells_in_mask]
-    dFF = dFF[cells_in_mask]
-    num_dff = dFF.shape[-1]
-
-    numCore = 450
-    cluster, client = fdask.setup_workers(numCore=numCore,is_local=False)
-    fdask.print_client_links(client)
-    print(client.dashboard_link)
-    if not os.path.exists(save_root+'cell_dff.zarr'):
-        dFF_ = zarr.array(dFF, chunks=(dFF.shape[0]//(numCore-2), dFF.shape[1]))
-        zarr.save(save_root+'cell_dff.zarr', dFF_)
-    dFF_ = da.from_zarr(save_root+'cell_dff.zarr')
-    clear_variables(dFF)
-
-    ###################################
-    ## Downsample sensory and motor input to frames
-    ###################################
-    ephys_dat = glob(ephys_dir+'/*.10chFlt')[0]
-    fileContent_ = load(ephys_dat)
-    l_power = windowed_variance(fileContent_[0])[0]
-    r_power = windowed_variance(fileContent_[1])[0]
-    camtrig = fileContent_[2]
-
-    expt_meta = glob(dat_dir+'ephys/*end*.xml')[0]
-    expt_paradigm = open_ephys_metadata(expt_meta)
-    probe_amp = (expt_paradigm.loc['LG probe']['velocity']*100).astype('int')
-    probe_gain = expt_paradigm.loc['LG probe']['gain']
-
-    indx = ep2frame(camtrig, thres=3.8)
-    frame_ = np.zeros(len(camtrig))
-    frame_[indx]=1
-    frame_ = frame_.cumsum()
-
-    slide_win = 180000
-    r_power_baseline = rolling_perc(r_power, window=slide_win, perc=0.1)
-    l_power_baseline = rolling_perc(l_power, window=slide_win, perc=0.1)
-
-    l_power_ = np.clip(l_power-l_power_baseline, 0, None)*10000
-    r_power_ = np.clip(r_power-r_power_baseline, 0, None)*10000
-
-    frame_len = np.min(np.unique(indx[1:]-indx[:-1]))
-
-    epoch_frame = np.median(wrap_data(fileContent_[5], indx, frame_len), axis=0).astype('int')[:num_dff]
-    swim_frame = np.mean(wrap_data(l_power_, indx, frame_len), axis=0)[:num_dff]
-    pulse_frame = np.rint(np.median(wrap_data(fileContent_[8], indx, frame_len), axis=0)*100).astype('int')[:num_dff]
-    visu_frame = np.mean(wrap_data(fileContent_[3], indx, frame_len), axis=0)[:num_dff]
-    visu_frame_ = visu_frame.copy()
-    visu_frame_[visu_frame_<0]=0
-    
-    ###################################
-    ## passive cells
-    ###################################
-    swim_thres = max(np.percentile(swim_frame, 85), 0.2)
-    active_pulse_trial = []
-    passive_pulse_trial = []
-    epoch_on = np.where((epoch_frame[1:]%5==0) & (epoch_frame[:-1]%5>0))[0]+1
-    epoch_on = np.r_[0, epoch_on, len(epoch_frame)]
-    t_pre = 5
-    t_post = 50
-    passive_trial = []
-    for n_ in range(len(epoch_on)-1):
-        on_ = epoch_on[n_]
-        off_ = epoch_on[n_+1]
-        swim_ = np.clip(swim_frame[on_:off_]-swim_thres, 0, np.inf)
-        epoch_ = epoch_frame[on_:off_]
-        pulse_ = pulse_frame[on_:off_]
-        if (epoch_%5==3).sum()<120:
-            continue
-        if swim_[epoch_%5==0].sum()==0:
-            continue
-        if swim_[epoch_%5==1].sum()==0:
-            continue
-        swim_len_ = np.where(swim_[epoch_%5==1]>0)[0]
-        swim_len_ = swim_len_.max() - swim_len_.min()
-        if swim_len_<10:
-            continue
-        if swim_[epoch_%5==2][5:].sum()>0:
-            plt.plot(swim_[epoch_%5==2])
-            plt.show()
-            continue
-        trial_type_ = epoch_[0]//5
-        pulse_on_ = np.where(pulse_==pulse_amp)[0]
-        if len(pulse_on_)==0: # skip catch trials
-            continue
-        pulse_on_ = pulse_on_[0]
-        if swim_[pulse_on_-t_pre:pulse_on_+t_post].sum()>0:
-            continue
-        passive_trial.append([trial_type_, pulse_on_+on_])
-    passive_trial=np.array(passive_trial).astype('int')
-
-    active_pulse_trial = passive_trial[passive_trial[:,0]==0, 1]
-    passive_pulse_trial = passive_trial[passive_trial[:,0]==1, 1]
-
-    if (len(active_pulse_trial)>0) and (len(passive_pulse_trial)>0):
-        cell_active_pulse_stats = dFF_.map_blocks(comp_stats_chunks, cond_trial=active_pulse_trial, comp_trial=passive_pulse_trial, pre=t_pre, post=t_post, dtype='O').compute()  
-        np.savez(save_root+'cell_active_pulse_stats', cell_active_pulse_stats=cell_active_pulse_stats)
-
-    fdask.terminate_workers(cluster, client)
-    return None
